requirements.txt
flask
jass-kit
sklearn
tensorflow